@@ -11,15 +11,16 @@
     ],
     "prefer-stable": true,
     "require": {
-        "php": "^7.2",
-        "symfony/framework-bundle": "^3.0|^4.0|^5.0",
-        "insidestyles/swoole-bridge": "^0.4.0",
-        "symfony/psr-http-message-bridge": "^1.0",
-        "symfony/console": "^3.0|^4.0|^5.0",
-        "zendframework/zend-diactoros": "^1.7"
+        "php": ">=7.2",
+        "symfony/framework-bundle": "^5.0|^6.0",
+        "insidestyles/swoole-bridge": "^1.0",
+        "symfony/psr-http-message-bridge": "^1.0|^2.0",
+        "symfony/console": "^5.0|^6.0",
+        "laminas/laminas-diactoros": "^2.0",
+        "nyholm/psr7": "^1.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.5",
+        "phpunit/phpunit": ">=6.5",
         "eaglewu/swoole-ide-helper": "dev-master"
     },
     "suggest": {
